+++ onnx/defs/doc_strings.cc
@@ -449,7 +449,7 @@ In more detail, the conversion among numerical types should follow these rules
 if the destination type is not a float 8 type.
 
 * Casting from floating point to:
-  * floating point: +/- infinity if OOR (out of range).
+  * floating point: round to nearest even (RNE) for in-range values; +/- infinity if OOR (out of range).
   * fixed point: undefined if OOR.
   * bool: +/- 0.0 to False; all else to True.
 * Casting from fixed point to: